+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Changed linters to mypy and isort
 ### Fixed
 * Added headers to httpx.AsyncClient, that supports content in body to be json
+* Setting.metadata default value is now an empty dict
 ## 0.1.2
 #### Fixed
 * Removed redundant warning from `subclasses.py` when using `TRACK_ALL` feature.